Drop Pydantic
Unfortunately, Pydantic is not native compatible with SQLAlchemy. The solution is use SLQModel from Sebastion. Although it'd add another dependency (I can live with that with all the trade-off known), the main reason is that SQLModel makes a reference to the models as SQL classes. This application uses DDD thus it is necessary to remove any kind of references to the DB from the domain classes.